refactor: LottoTicket을 생성하는 로직 LottoTicketGenerator로 이관
